feat: ファイルベースルーティングに移行
- app/page.tsx をアップロード画面専用に整理（1028行→80行）
- app/projects/page.tsx でプロジェクト一覧
- app/projects/[id]/page.tsx で詳細画面（動的ルート）
- app/settings/page.tsx で設定画面
- app/contexts/AppContext.tsx で状態共有
- app/components/AppHeader.tsx, AppSnackbar.tsx を共通化
- layout.tsx で全ページ共通レイアウト
