+++ app/layout.tsx
@@ -1,5 +1,8 @@
 import type { Metadata } from "next";
 import "./globals.css";
+import { AppProvider } from "./contexts/AppContext";
+import { AppHeader } from "./components/AppHeader";
+import { AppSnackbar } from "./components/AppSnackbar";
 
 export const metadata: Metadata = {
   title: "SBOM Vulnerability Scanner",
@@ -14,7 +17,11 @@ export default function RootLayout({
   return (
     <html lang="ja">
       <body style={{ margin: 0, padding: 0 }}>
-        {children}
+        <AppProvider>
+          <AppHeader />
+          <main>{children}</main>
+          <AppSnackbar />
+        </AppProvider>
       </body>
     </html>
   );
